[RFC][DO NOT MERGE] sched/uclamp: Filter out uclamp_{max, min} for small tasks
DISCLAIMER:
=====================================================================

This patch is intended to go upstream after collecting feedback from
Android community that it resolves the issues reported by various
partners. It is not meant to be merged into android-mainline.

Please test the patch and let us know how well it works for your use
cases.

=====================================================================

Changes in v2:<<<<<<< HEAD

	rq->uclamp_flags = UCLAMP_FLAG_IDLE;

- Remove uclamp_rq_inc() and instead do the call from
  enqueue_task_{fair, rt}(). This helps us to avoid the unnecessary
  uclamp_rq_inc_id() followed by uclamp_rq_dec_id() by the filter.

- Remove uclamp_min check from uclamp_can_ignore_uclamp_max()
  Many transient task in top-app group could be boosted and the check
  was overzealous. So remove it.

- Runtime info are not reliable when the task is migrating. So check for
  that in uclamp_can_ignore_uclamp_max().
  The problem is that active migration could happen after we ran for
  a brief amount of time. The runtime does not then represent the
  duration of the previous wake up.

- Change the divider to be a simple number instead of
	divider = 1 << sched_util_clamp_max_filter_divider

- Introduce a new uclamp_can_ignore_uclamp_min()

  Wei raised that with inability to do better classification than
  what can be done on cgroup ATM, any small tasks that don't run long
  enough could still generate requests to be boosted (as set by
  top-app->cpu.uclamp.min cgroup), which could end wasting unnecessary
  power.

  We implement a logic to ignore uclamp_min in a similar fashion to
  uclamp_max such that task runtime must be sufficiently long to honour
  the boost:

    runtime > policy->cpuinfo.transition_latency * multiplier

  where multiplier is sysctl_sched_util_clamp_min_filter_multiplier
  and can take a value of 0-16.

  The need for a uclamp_min filter is still debatable. Userspace being
  aggressive in boosting these small tasks is the root cause of the
  problem.

- Introduce new sysctl_sched_util_clamp_min_filter_multiplier to be used
  in conjunction with uclamp_can_ignore_uclamp_min()

- Address potential race condition where we could end up updating
  schedutil several times in the enqueue_task_fair() path.
  update_load_avg() can trigger an update, as well as our call to
  cpufreq_update_util() after uclamp_rq_inc_id() calls. We add a special
  flag to ensure we skip this call from enqueue_task_fair() path. We
  instead do a single call after all necessary processing was done.

- Refactor UCLAMP_FLAG_IDLE to behave better.
  Since we now can do uclamp_rq_{inc, dec}_id() with filtering, we can
  easily end up with imbalance and the current implementation wasn't
  designed to deal with multiple callers from different call site.

  Refactor the code so that we have individual flag for UCLAMP_MIN and
  UCLAMP_MAX to tag that the current rq value is the idle value. Which
  allows us to handle the logic completely within
  uclamp_rq_{inc, dec}_id() logic without relying on the callers doing
  the right thing which was fragile anyway. As we've seen when we
  introduced uclamp_rq_reinc_id().

WARNING:
========================================================================

The code that looks at policy->cpuinfo.transition_latency is
a bit fragile at the moment if the cpufreq governor changes a lot during
runtime. Which I hope is not the case for normal operation. This will be
addressed in later versions.

========================================================================

Origiinal commit message:
=========================

uclamp_max effectiveness could be easily impacted by small transient
tasks that wake up frequency to do small work then go back to sleep.

If there's a busy task that is capped by uclamp_max to run at a smaller
frequency, due to max-aggregation rule tasks that wake up on the same
cpu will increase the rq->uclamp_max value if they were higher than the
capped task. Given that all tasks by default have a uclamp_max = 1024,
this is the likely case by default.

Note that since the capped task is likely to be a busy and throttled
one, its util, and hence the rq->util, will be very high and as soon as
we lift the capping the requested frequency will be very high.

To address this issue of increasing the resilience of uclamp_max against
these transient tasks that don't really need to run at a higher
frequency, we implement a simple filter mechanism to ignore uclamp_max
for those tasks.

The algorithm looks at the runtime of the task and compares it to
sched_slice(). By default we assume any task that its runtime is 1/4th
of sched_slice() or less is a small transient task that we can ignore
its uclamp_max requirement.

	runtime < sched_slice() / divider

We can tweak the divider by
/proc/sys/kernel/sched_util_uclamp_max_filter_divider sysctl. It accepts
values 1-16.

We add a new task_tick_uclamp() function to verify this condition
periodically and ensure the conditions checked at wake up are still true
- in case this transient task suddenly becomes a busy one.

For EAS, we can't use sched_slice() there to figure out if uclamp_max
will be ignored because the task is not enqueued yet. So we leave it
as-is to figure out the placement based on worst case scenario.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>
Change-Id: Ie3afa93a7d70dab5b7c22e820cc078ffd0e891ef
Signed-off-by: acuicultor <ramoncuenca4@gmail.com>